utilities/requirements.txt
pyYAML~=6.0
pip~=21.1.2
setuptools~=57.0.0
wheel~=0.36.2